requirements.txt
PyMuPDF==1.26.4
Pillow==11.3.0
pytesseract==0.3.13
gtts==2.5.4
pydub==0.25.1
audioop-lts==0.2.2
pyttsx3==2.99
